v0.0.49

+++ server/api/rotina.js
@@ -10,9 +10,27 @@ var _params = require('common/sequelize/params');
 
 var _params2 = _interopRequireDefault(_params);
 
+var _usuario = require('common/models/portal/usuario');
+
+var _usuario2 = _interopRequireDefault(_usuario);
+
 function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
 
 module.exports = router => {
+
+    router.get('/rotina/:id', async (req, res, next) => {
+        const { RotinaInfor } = _sequelize2.default.models;
+        const { id } = req.params;
+        try {
+            const response = await RotinaInfor.scope('paineis', 'tags', 'responsaveis', 'ferramentas', 'periodos', 'usuarioInclusao').findById(id, {
+                order: _sequelize2.default.literal('"responsaveis->RotinaResponsavel"."tipo_id", "paineis"."nome"')
+
+            });
+            res.send(response);
+        } catch (err) {
+            next(err);
+        }
+    });
 
     router.post('/', async (req, res, next) => {
         const { RotinaInfor, RotinaPainel, RotinaTag, RotinaResponsavel, RotinaFerramenta, PeriodoRotina } = _sequelize2.default.models;
@@ -51,7 +69,7 @@ module.exports = router => {
                 await RotinaFerramenta.destroy({ where: { rotina_id: response.id } });
                 Promise.all(data.ferramentas.map(ferramenta_id => RotinaFerramenta.build({ rotina_id: response.id, ferramenta_id }).save()));
             }
-            if (data.periodos.length) {
+            if (data.periodos && data.periodos.length) {
                 await PeriodoRotina.destroy({ where: { rotina_id: response.id } });
                 Promise.all(data.periodos.map(periodo => PeriodoRotina.build(_extends({ rotina_id: response.id }, periodo)).save()));
             }
